pyproject.toml
[tool.poetry]
name = "zapbot"
version = "0.1.0"
description = ""
authors = ["MieleSantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
websockets = "^12.0"
aiohttp = "^3.9.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.0"
fastapi = {extras = ["standard"], version = "^0.117.1"}
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
langchain-experimental = "^0.3.4"
langchain-openai = "^0.3.33"
duckduckgo-search = "^8.1.1"
langchain-groq = "^0.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
